Simplifies preprocess_mpas* to one function
This commit simplifies the code by merging
`preprocess_mpas_timeSeriesStats` and `preprocess_mpas`, which were
essentially identical functions. The functions are combined, e.g., via a
flag like `timeSeriesStats=True` in a single `preprocess_mpas` function.

Note, however, that yearoffset=1849 for preprocess_mpas which
is a deviation from the previous version of the preprocessor.